src: create include function for file sourcing
Sourcing files with `. file.sh --sources-only` has a few issues:
the readibility is low and there is no mechanism to prevent a file from
being sourced twice.

With this approach, we solve all of this problems and we are also able
to create a unit test for our include function

This commit creates the include function and makes the entire codebase use
it whenever a file needs to be sourced (except for setup.sh).

Closes #197

remote: make it impossible to import remote before kw_conf_loader

docs: add file sourcing to codingstyle.rst

tests: add test for kw_include.sh

Signed-off-by: alan-barzilay <alan.barzilay@gmail.com>
Signed-off-by: João Seckler <jseckler@riseup.net>